backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
numpy==2.0.1
opencv-python==4.10.0.84
rembg==2.0.61
onnxruntime==1.18.1